Added symbols file for libarchive.  Closes: #476516.
Attached is a symbols file for libarchive. Just put it in debian/ and
dh_makeshlibs will call dpkg-gensymbols at build time.  This will
cause programs linking against this library to have less strict
dependencies, so they can migrate to testing (or in some future
perhaps even be installed on stable without recompiling) when they
only need old functionality.

When you put the file there, it will need to be kept up to date:
dpkg-gensymbols will automatically output a diff and update the file
if any new symbols got added. What it cannot detect if is some
function allows more magic values as parameters than before. If that
is ever the case (I could not find any indication of this since
2.0.25) the function will have to be increased manually (unless the
new constant is not useable without using some other new function,
then that will already pull the dependencies high enough).

Thanks in advance,
       Bernhard R. Link

I've only tested this on one architecture (sparc), but given that it
is a C library, I doubt there will be any architecture specific
symbols in there, which could complicate the situation...